requirements.txt
ollama>=0.3.0
qdrant-client>=1.9.0
langchain>=0.2.0
langchain-community>=0.2.0
sentence-transformers>=3.0.0
numpy>=1.24.0
pandas>=2.0.0
python-dotenv>=1.0.0
pypdf>=4.0.0
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6
aiofiles>=23.0.0